+++ hosts/matangi/default.nix
@@ -2,6 +2,7 @@
   lib,
   inputs,
   outputs,
+  pkgs,
   ...
 }: {
   imports = [
@@ -226,6 +227,8 @@
   };
 
   hardware.nvidia.forceFullCompositionPipeline = lib.mkForce false;
+
+  environment.systemPackages = [ pkgs.easyeffects ];
 
   services = {
     logind.settings.Login = {