+++ modules/nixos/desktop/xfce.nix
@@ -1,23 +1,19 @@
 {
   config,
   lib,
-  pkgs,
   ...
 }:
 with lib; {
   options.modules.desktop.xfce.enable = mkEnableOption "XFCE desktop environment";
+
   config = mkIf config.modules.desktop.xfce.enable {
     services.xserver = {
       enable = true;
       desktopManager.xfce.enable = true;
     };
+
     services.devmon.enable = true;
     services.udisks2.enable = true;
     services.acpid.enable = true;
-
-    environment.variables = {
-      GDK_BACKEND = "x11";
-      QT_QPA_PLATFORM = "xcb";
-    };
   };
 }